+++ lapce-core/src/buffer/rope_text.rs
@@ -3,7 +3,10 @@ use std::{borrow::Cow, ops::Range};
 use lsp_types::Position;
 use xi_rope::{interval::IntervalBounds, Cursor, Rope};
 
-use crate::encoding::{offset_utf16_to_utf8, offset_utf8_to_utf16};
+use crate::{
+    encoding::{offset_utf16_to_utf8, offset_utf8_to_utf16},
+    word::WordCursor,
+};
 
 pub struct RopeText<'a> {
     text: &'a Rope,
@@ -162,6 +165,25 @@ impl<'a> RopeText<'a> {
             }
         }
         new_offset
+    }
+
+    pub fn first_non_blank_character_on_line(&self, line: usize) -> usize {
+        let last_line = self.last_line();
+        let line = if line > last_line + 1 {
+            last_line
+        } else {
+            line
+        };
+        let line_start_offset = self.text.offset_of_line(line);
+        WordCursor::new(&self.text, line_start_offset).next_non_blank_char()
+    }
+
+    pub fn indent_on_line(&self, line: usize) -> String {
+        let line_start_offset = self.text.offset_of_line(line);
+        let word_boundary =
+            WordCursor::new(&self.text, line_start_offset).next_non_blank_char();
+        let indent = self.text.slice_to_cow(line_start_offset..word_boundary);
+        indent.to_string()
     }
 
     pub fn slice_to_cow(&self, range: Range<usize>) -> Cow<'a, str> {
